DOCUMENTATION: Comm Lab Assignment 1
Description:
Create a web page to house and debut your 30MFF. Think about how you want to "brand" the film.

Project Description: 
My project talks about an honest website for an honest zoom review !! When we first made the film as a group, I thought what would it be like to have an honest website for the same:) Therefore, this a simple website but with honest structure like the main page description and the area where "big word" describe the movie as they do in vlog websites when some top rank company reviews them.  The website then also show some random engagement rate and subscribers. 
The second page is about the team which I felt had to be talked about . I wrote a short description about our work and how we completed it in 45 mins over "zoom". The picture next to it is the only picture the four of us have. 
The end page where we have the movie is the area that involves a bit activity as well. There are buttons on the bottom to pause/play the movie and choose the screen size or the preferred size of display. 
Overall the website acts like an application like zoom with navigation panel on the side instead on top like normal mobile applications( this is not compulsory but it was just an observation from companies like facebook and delivery companies like careem who have their navigations on the side instead of top).

Process: 
I started with the colour theme of the website which is very similar to that of zoom: blue and grey/white. Then I created the navigation panel on the side where I added theme like home page, about us and "ZOOM: the movie". Then I started by putting small elements on each page starting with the home page. 
The home page has a small welcome message and an "honest" application of that welcome message and next to that is a picture of zoom. Then as we scroll down we come across the words describing the page and all the boxes. All the text were created using the normal text on html and text  defined background styles on css. 
Similar to the home page the about us page, using the same style talks about the team and how they created the movie. 
Then on the movie page, we see the movie and using  java script, we can see that the movie screen in adjustable. I created the video as a variable, and then created functions for each button for the variable(video). 

Challenges: 
It was difficult at first to get javascript and html right for the website with respect to positioning. In fact positioning a s a whole was very challenging since I used html for the first time in my life. Overall, it was quite an enriching experience and  I am more confident about my coding skills than I was before.